+++ tests/conftest.py
@@ -51,7 +51,12 @@ def current_client_token(admin_client: DynamicClient) -> str:
 
 @pytest.fixture(scope="class")
 def model_namespace(request: FixtureRequest, admin_client: DynamicClient) -> Generator[Namespace, Any, Any]:
-    with create_ns(admin_client=admin_client, name=request.param["name"]) as ns:
+    ns_kwargs = {"admin_client": admin_client, "name": request.param["name"]}
+
+    if request.param.get("modelmesh-enabled"):
+        ns_kwargs["labels"] = {"modelmesh-enabled": "true"}
+
+    with create_ns(**ns_kwargs) as ns:
         yield ns
 
 
@@ -167,16 +172,6 @@ def vllm_runtime_image(pytestconfig: pytest.Config) -> str | None:
     if not runtime_image:
         return None
     return runtime_image
-
-
-@pytest.fixture(scope="class")
-def ns_with_modelmesh_enabled(request: FixtureRequest, admin_client: DynamicClient) -> Generator[Namespace, Any, Any]:
-    with create_ns(
-        admin_client=admin_client,
-        name=request.param["name"],
-        labels={"modelmesh-enabled": "true"},
-    ) as ns:
-        yield ns
 
 
 @pytest.fixture(scope="session")
