+++ tests/model_serving/model_server/conftest.py
@@ -17,7 +17,17 @@ from ocp_resources.storage_class import StorageClass
 from pytest_testconfig import config as py_config
 
 from tests.model_serving.model_server.utils import create_isvc
-from utilities.constants import DscComponents, StorageClassName
+from utilities.constants import (
+    DscComponents,
+    KServeDeploymentType,
+    ModelAndFormat,
+    ModelFormat,
+    ModelInferenceRuntime,
+    ModelVersion,
+    Protocols,
+    RuntimeTemplates,
+    StorageClassName,
+)
 from utilities.infra import s3_endpoint_secret
 from utilities.data_science_cluster_utils import update_components_in_dsc
 from utilities.serving_runtime import ServingRuntimeFromTemplate
@@ -208,3 +218,190 @@ def skip_if_no_deployed_openshift_service_mesh(admin_client: DynamicClient):
         pytest.skip("OpenShift service mesh operator is not deployed")
 
     smcp.wait_for_condition(condition=smcp.Condition.READY, status="True")
+
+
+@pytest.fixture(scope="class")
+def http_s3_openvino_model_mesh_inference_service(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    http_s3_ovms_model_mesh_serving_runtime: ServingRuntime,
+    ci_model_mesh_endpoint_s3_secret: Secret,
+    model_mesh_model_service_account: ServiceAccount,
+) -> InferenceService:
+    with create_isvc(
+        client=admin_client,
+        name=f"{Protocols.HTTP}-{ModelFormat.OPENVINO}",
+        namespace=model_namespace.name,
+        runtime=http_s3_ovms_model_mesh_serving_runtime.name,
+        model_service_account=model_mesh_model_service_account.name,
+        storage_key=ci_model_mesh_endpoint_s3_secret.name,
+        storage_path=request.param["model-path"],
+        model_format=ModelAndFormat.OPENVINO_IR,
+        deployment_mode=KServeDeploymentType.MODEL_MESH,
+        model_version=ModelVersion.OPSET1,
+    ) as isvc:
+        yield isvc
+
+
+@pytest.fixture(scope="class")
+def http_s3_ovms_model_mesh_serving_runtime(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+) -> ServingRuntime:
+    rt_kwargs = {
+        "client": admin_client,
+        "namespace": model_namespace.name,
+        "name": f"{Protocols.HTTP}-{ModelInferenceRuntime.OPENVINO_RUNTIME}",
+        "template_name": RuntimeTemplates.OVMS_MODEL_MESH,
+        "multi_model": True,
+        "protocol": "REST",
+        "resources": {
+            "ovms": {
+                "requests": {"cpu": "1", "memory": "4Gi"},
+                "limits": {"cpu": "2", "memory": "8Gi"},
+            }
+        },
+    }
+
+    enable_external_route = False
+    enable_auth = False
+
+    if hasattr(request, "param"):
+        enable_external_route = request.param.get("enable-external-route")
+        enable_auth = request.param.get("enable-auth")
+
+    rt_kwargs["enable_external_route"] = enable_external_route
+    rt_kwargs["enable_auth"] = enable_auth
+
+    with ServingRuntimeFromTemplate(**rt_kwargs) as model_runtime:
+        yield model_runtime
+
+
+@pytest.fixture(scope="class")
+def ci_model_mesh_endpoint_s3_secret(
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    aws_access_key_id: str,
+    aws_secret_access_key: str,
+    ci_s3_bucket_name: str,
+    ci_s3_bucket_region: str,
+    ci_s3_bucket_endpoint: str,
+) -> Secret:
+    with s3_endpoint_secret(
+        admin_client=admin_client,
+        name="ci-bucket-secret",
+        namespace=model_namespace.name,
+        aws_access_key=aws_access_key_id,
+        aws_secret_access_key=aws_secret_access_key,
+        aws_s3_region=ci_s3_bucket_region,
+        aws_s3_bucket=ci_s3_bucket_name,
+        aws_s3_endpoint=ci_s3_bucket_endpoint,
+    ) as secret:
+        yield secret
+
+
+@pytest.fixture(scope="class")
+def model_mesh_model_service_account(
+    admin_client: DynamicClient, ci_model_mesh_endpoint_s3_secret: Secret
+) -> ServiceAccount:
+    with ServiceAccount(
+        client=admin_client,
+        namespace=ci_model_mesh_endpoint_s3_secret.namespace,
+        name=f"{Protocols.HTTP}-models-bucket-sa",
+        secrets=[{"name": ci_model_mesh_endpoint_s3_secret.name}],
+    ) as sa:
+        yield sa
+
+
+@pytest.fixture(scope="class")
+def openvino_kserve_serving_runtime(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+) -> ServingRuntime:
+    with ServingRuntimeFromTemplate(
+        client=admin_client,
+        name=request.param["runtime-name"],
+        namespace=model_namespace.name,
+        template_name=RuntimeTemplates.OVMS_KSERVE,
+        multi_model=False,
+        resources={
+            "ovms": {
+                "requests": {"cpu": "1", "memory": "4Gi"},
+                "limits": {"cpu": "2", "memory": "8Gi"},
+            }
+        },
+        model_format_name=request.param["model-format"],
+    ) as model_runtime:
+        yield model_runtime
+
+
+@pytest.fixture(scope="class")
+def ci_endpoint_s3_secret(
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    aws_access_key_id: str,
+    aws_secret_access_key: str,
+    ci_s3_bucket_name: str,
+    ci_s3_bucket_region: str,
+    ci_s3_bucket_endpoint: str,
+) -> Secret:
+    with s3_endpoint_secret(
+        admin_client=admin_client,
+        name="ci-bucket-secret",
+        namespace=model_namespace.name,
+        aws_access_key=aws_access_key_id,
+        aws_secret_access_key=aws_secret_access_key,
+        aws_s3_region=ci_s3_bucket_region,
+        aws_s3_bucket=ci_s3_bucket_name,
+        aws_s3_endpoint=ci_s3_bucket_endpoint,
+    ) as secret:
+        yield secret
+
+
+@pytest.fixture(scope="class")
+def ovms_serverless_inference_service(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    openvino_kserve_serving_runtime: ServingRuntime,
+    ci_endpoint_s3_secret: Secret,
+) -> InferenceService:
+    with create_isvc(
+        client=admin_client,
+        name=f"{request.param['name']}-serverless",
+        namespace=model_namespace.name,
+        runtime=openvino_kserve_serving_runtime.name,
+        storage_path=request.param["model-dir"],
+        storage_key=ci_endpoint_s3_secret.name,
+        model_format=ModelAndFormat.OPENVINO_IR,
+        deployment_mode=KServeDeploymentType.SERVERLESS,
+        model_version=request.param["model-version"],
+    ) as isvc:
+        yield isvc
+
+
+@pytest.fixture(scope="class")
+def http_s3_tensorflow_model_mesh_inference_service(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    http_s3_ovms_model_mesh_serving_runtime: ServingRuntime,
+    ci_model_mesh_endpoint_s3_secret: Secret,
+    model_mesh_model_service_account: ServiceAccount,
+) -> InferenceService:
+    with create_isvc(
+        client=admin_client,
+        name=f"{Protocols.HTTP}-{ModelFormat.TENSORFLOW}",
+        namespace=model_namespace.name,
+        runtime=http_s3_ovms_model_mesh_serving_runtime.name,
+        model_service_account=model_mesh_model_service_account.name,
+        storage_key=ci_model_mesh_endpoint_s3_secret.name,
+        storage_path=request.param["model-path"],
+        model_format=ModelFormat.TENSORFLOW,
+        deployment_mode=KServeDeploymentType.MODEL_MESH,
+        model_version="2",
+    ) as isvc:
+        yield isvc
