+++ tests/model_serving/model_server/model_mesh/test_model_mesh_multi_models.py
@@ -8,39 +8,26 @@ from utilities.constants import (
 )
 from utilities.inference_utils import Inference
 
-
 pytestmark = [pytest.mark.modelmesh]
 
 
 @pytest.mark.parametrize(
-    "model_namespace, http_s3_openvino_model_mesh_inference_service",
+    "model_namespace, http_s3_openvino_model_mesh_inference_service, "
+    "http_s3_openvino_second_model_mesh_inference_service",
     [
         pytest.param(
-            {"name": "model-mesh-openvino", "modelmesh-enabled": True},
+            {"name": "model-mesh-multi-model", "modelmesh-enabled": True},
             {"model-path": ModelStoragePath.OPENVINO_EXAMPLE_MODEL},
+            {"model-path": ModelStoragePath.OPENVINO_VEHICLE_DETECTION},
         )
     ],
     indirect=True,
 )
-class TestOpenVINOModelMesh:
-    @pytest.mark.smoke
-    @pytest.mark.polarion("ODS-2053", "ODS-2054")
-    def test_model_mesh_openvino_rest_inference_internal_route(self, http_s3_openvino_model_mesh_inference_service):
-        verify_inference_response(
-            inference_service=http_s3_openvino_model_mesh_inference_service,
-            runtime=ModelInferenceRuntime.OPENVINO_RUNTIME,
-            inference_type=Inference.INFER,
-            protocol=Protocols.HTTP,
-            use_default_query=True,
-        )
-
-    @pytest.mark.sanity
-    @pytest.mark.polarion("ODS-1920")
-    def test_model_mesh_openvino_inference_with_token(
+class TestOpenVINOModelMeshMultiModels:
+    def test_model_mesh_openvino_inference_with_tensorflow(
         self,
-        patched_model_mesh_sr_with_authentication,
         http_s3_openvino_model_mesh_inference_service,
-        model_mesh_inference_token,
+        http_s3_openvino_second_model_mesh_inference_service,
     ):
         verify_inference_response(
             inference_service=http_s3_openvino_model_mesh_inference_service,
@@ -48,5 +35,17 @@ class TestOpenVINOModelMesh:
             inference_type=Inference.INFER,
             protocol=Protocols.HTTP,
             use_default_query=True,
-            token=model_mesh_inference_token,
+        )
+
+    def test_model_mesh_tensorflow_with_openvino(
+        self,
+        http_s3_openvino_second_model_mesh_inference_service,
+        http_s3_openvino_model_mesh_inference_service,
+    ):
+        verify_inference_response(
+            inference_service=http_s3_openvino_second_model_mesh_inference_service,
+            runtime=ModelInferenceRuntime.TENSORFLOW_RUNTIME,
+            inference_type=Inference.INFER,
+            protocol=Protocols.HTTP,
+            query_input="@utilities/manifests/openvino/vehicle-detection-inputs.txt",
         )
