--- tests/model_serving/model_server/ovms/kserve/conftest.py
@@ -1,84 +1,0 @@
-import pytest
-from _pytest.fixtures import FixtureRequest
-from kubernetes.dynamic import DynamicClient
-from ocp_resources.inference_service import InferenceService
-from ocp_resources.namespace import Namespace
-from ocp_resources.secret import Secret
-from ocp_resources.serving_runtime import ServingRuntime
-
-from tests.model_serving.model_server.utils import create_isvc
-from utilities.constants import (
-    KServeDeploymentType,
-    ModelAndFormat,
-    RuntimeTemplates,
-)
-from utilities.infra import s3_endpoint_secret
-from utilities.serving_runtime import ServingRuntimeFromTemplate
-
-
-@pytest.fixture(scope="class")
-def openvino_kserve_serving_runtime(
-    request: FixtureRequest,
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-) -> ServingRuntime:
-    with ServingRuntimeFromTemplate(
-        client=admin_client,
-        name=request.param["runtime-name"],
-        namespace=model_namespace.name,
-        template_name=RuntimeTemplates.OVMS_KSERVE,
-        multi_model=False,
-        resources={
-            "ovms": {
-                "requests": {"cpu": "1", "memory": "4Gi"},
-                "limits": {"cpu": "2", "memory": "8Gi"},
-            }
-        },
-        model_format_name=request.param["model-format"],
-    ) as model_runtime:
-        yield model_runtime
-
-
-@pytest.fixture(scope="class")
-def ci_endpoint_s3_secret(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    aws_access_key_id: str,
-    aws_secret_access_key: str,
-    ci_s3_bucket_name: str,
-    ci_s3_bucket_region: str,
-    ci_s3_bucket_endpoint: str,
-) -> Secret:
-    with s3_endpoint_secret(
-        admin_client=admin_client,
-        name="ci-bucket-secret",
-        namespace=model_namespace.name,
-        aws_access_key=aws_access_key_id,
-        aws_secret_access_key=aws_secret_access_key,
-        aws_s3_region=ci_s3_bucket_region,
-        aws_s3_bucket=ci_s3_bucket_name,
-        aws_s3_endpoint=ci_s3_bucket_endpoint,
-    ) as secret:
-        yield secret
-
-
-@pytest.fixture(scope="class")
-def ovms_serverless_inference_service(
-    request: FixtureRequest,
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    openvino_kserve_serving_runtime: ServingRuntime,
-    ci_endpoint_s3_secret: Secret,
-) -> InferenceService:
-    with create_isvc(
-        client=admin_client,
-        name=f"{request.param['name']}-serverless",
-        namespace=model_namespace.name,
-        runtime=openvino_kserve_serving_runtime.name,
-        storage_path=request.param["model-dir"],
-        storage_key=ci_endpoint_s3_secret.name,
-        model_format=ModelAndFormat.OPENVINO_IR,
-        deployment_mode=KServeDeploymentType.SERVERLESS,
-        model_version=request.param["model-version"],
-    ) as isvc:
-        yield isvc
