+++ tests/model_serving/model_server/ovms/model_mesh/test_tensorflow.py
@@ -13,11 +13,10 @@ pytestmark = [pytest.mark.modelmesh]
 
 
 @pytest.mark.parametrize(
-    "ns_with_modelmesh_enabled, http_s3_ovms_model_mesh_serving_runtime,"
-    " http_s3_tensorflow_model_mesh_inference_service",
+    "model_namespace, http_s3_ovms_model_mesh_serving_runtime, http_s3_tensorflow_model_mesh_inference_service",
     [
         pytest.param(
-            {"name": "model-mesh-tensorflow"},
+            {"name": "model-mesh-tensorflow", "modelmesh-enabled": True},
             {"enable-external-route": True},
             {"model-path": ModelStoragePath.TENSORFLOW_MODEL},
         )
